20200620 - Merge

+++ piratebox/piratebox/bin/install_piratebox.sh
@@ -1,13 +1,23 @@
 #!/bin/sh
-## PirateBox installer script  v.01
-##  created by Matthias Strubel   2011-08-04
+## PirateBox installer script  
+##  created by Matthias Strubel   (c)2011-2014 GPL-3
 ##
 
-## ASH does not support arrays, so no nice foreach 
-# All Perl packages for kareha
-##OPENWRT_PACKAGES_IMAGEBOARD=(  perl perlbase-base perlbase-cgi perlbase-essential perlbase-file perlbase-bytes perlbase-config perlbase-data perlbase-db-file perlbase-digest perlbase-encode perlbase-encoding perlbase-fcntl perlbase-gdbm-file perlbase-integer perlbase-socket perlbase-unicode perlbase-utf8 perlbase-xsloader  )
+create_content_folder(){
 
+   echo "Creating 'content' folder on USB stick and move over stuff"
+   mkdir -p $WWW_CONTENT
+   cp -r     $PIRATEBOX_FOLDER/www_content/*   $WWW_CONTENT
 
+   [ ! -L $PIRATEBOX_FOLDER/www/content  ] && \
+		ln -s $WWW_CONTENT  $WWW_FOLDER/content
+   [ ! -e $WWW_FOLDER/favicon.ico ] && \
+		ln -s $WWW_CONTENT/favicon.ico $WWW_FOLDER
+
+   chown $LIGHTTPD_USER:$LIGHTTPD_GROUP  $WWW_CONTENT -R
+   chmod  u+rw $WWW_CONTENT
+   return 0
+}
 
 # Load configfile
 
@@ -36,13 +46,11 @@ PIRATEBOX_CONFIG=$1
 
 if [ $2 = 'pyForum' ] ; then
     cp -v $PIRATEBOX_FOLDER/src/forest.py  $WWW_FOLDER/cgi-bin
-    cp -v $PIRATEBOX_FOLDER/src/forest.css $WWW_FOLDER/
-    cp -v $PIRATEBOX_FOLDER/src/forum_forest.html  $WWW_FOLDER/forum.html
+    cp -v $PIRATEBOX_FOLDER/src/forest.css $WWW_FOLDER/content/css
     mkdir -p $PIRATEBOX_FOLDER/forumspace
-    chmod a+rw -R  $PIRATEBOX_FOLDER/forumspace
+    chmod a+rw -R  $PIRATEBOX_FOLDER/forumspace 2> /dev/null
     chown $LIGHTTPD_USER:$LIGHTTPD_GROUP  $WWW_FOLDER/cgi-bin/forest.py
-    chown $LIGHTTPD_USER:$LIGHTTPD_GROUP  $WWW_FOLDER/forest.css
-    chown $LIGHTTPD_USER:$LIGHTTPD_GROUP  $WWW_FOLDER/forum.html
+    chown $LIGHTTPD_USER:$LIGHTTPD_GROUP  $WWW_FOLDER/content/forest.css  2> /dev/null
     echo "Copied the files. Recheck your PirateBox now. "
 fi
 
@@ -57,10 +65,11 @@ if [ $2 = 'part2' ] ; then
    mkdir -p $PIRATEBOX_FOLDER/share/tmp
    mkdir -p $PIRATEBOX_FOLDER/tmp
 
-#Copy Forban-Link spacer
-   cp $PIRATEBOX_FOLDER/src/no_link.html $PIRATEBOX_FOLDER/www/forban_link.html
-
-#Set permissions
+   #Distribute the Directory Listing files
+   if [ "$CUSTOM_DIRLIST_COPY" = "yes" ] ; then
+       $PIRATEBOX_FOLDER/bin/distribute_files.sh $SHARE_FOLDER/Shared true
+   fi
+   #Set permissions
    chown $LIGHTTPD_USER:$LIGHTTPD_GROUP  $PIRATEBOX_FOLDER/share -R
    chmod  u+rw $PIRATEBOX_FOLDER/share
    chown $LIGHTTPD_USER:$LIGHTTPD_GROUP  $PIRATEBOX_FOLDER/www -R
@@ -73,18 +82,17 @@ if [ $2 = 'part2' ] ; then
    if  [ !  -f $PIRATEBOX_FOLDER/share/board/kareha.pl ] ; then  
       cp $PIRATEBOX_FOLDER/src/kareha.pl $PIRATEBOX_FOLDER/share/board
    fi
-   
-   ln -s $PIRATEBOX_FOLDER/share/board $PIRATEBOX_FOLDER/www/board
-   ln -s $UPLOADFOLDER  $PIRATEBOX_FOLDER/www/Shared
+  
+   [ ! -L $PIRATEBOX_FOLDER/www/board  ] && ln -s $PIRATEBOX_FOLDER/share/board $PIRATEBOX_FOLDER/www/board
+   [ ! -L $PIRATEBOX_FOLDER/www/Shared ] && ln -s $UPLOADFOLDER  $PIRATEBOX_FOLDER/www/Shared
+   [ ! -L $PIRATEBOX_FOLDER/www/content  ] && \
+       ln -s $WWW_CONTENT  $WWW_FOLDER/content
+
 fi 
 
 #Install the image-board
 if [ $2 = 'imageboard' ] ; then
    
-    #Activate on mainpage
-    cp $PIRATEBOX_FOLDER/src/forum_kareha.html  $WWW_FOLDER/forum.html
-
-
     if [ -e  $PIRATEBOX_FOLDER/share/board/init_done ] ; then
        echo "$PIRATEBOX_FOLDER/share/board/init_done file Found in Kareha folder. Won't reinstall board."
        exit 0;
@@ -95,7 +103,7 @@ if [ $2 = 'imageboard' ] ; then
     KAREHA_RELEASE=kareha_3.1.4.zip
     if [ ! -e $PIRATEBOX_FOLDER/tmp/$KAREHA_RELEASE ] ; then
 	echo "  Wgetting kareha-zip file "
-    	wget http://wakaba.c3.cx/releases/$KAREHA_RELEASE
+    	wget http://wakaba.c3.cx/releases/Kareha/$KAREHA_RELEASE
 	if [ "$?" != "0" ] ; then
        		echo "wget kareha failed.. you can place the current file your to  $PIRATEBOX_FOLDER/tmp "
 	 fi
@@ -115,12 +123,10 @@ if [ $2 = 'imageboard' ] ; then
     cd  $PIRATEBOX_FOLDER/share/board  
     cp -R  mode_image/* ./   
     cp  $PIRATEBOX_FOLDER/src/kareha_img_config.pl $PIRATEBOX_FOLDER/share/board/config.pl 
+    cp  $PIRATEBOX_FOLDER/src/no_forum.html  $PIRATEBOX_FOLDER/share/board/index.htm
     chown -R $LIGHTTPD_USER:$LIGHTTPD_GROUP  $PIRATEBOX_FOLDER/share/board   
     #Install filetype thumbnails
     mv $PIRATEBOX_FOLDER/share/board/extras/icons  $PIRATEBOX_FOLDER/share/board/ 
-
-    #Activate on mainpage
-    cp $PIRATEBOX_FOLDER/src/forum_kareha.html  $WWW_FOLDER/forum.html
 
     echo "Errors in chown occurs if you are using vfat on the USB stick"
     echo "   . don't Panic!"
@@ -157,7 +163,7 @@ fi
 set_hostname() {
 	local name=$1 ; shift;
 
-	sed  "s|#####HOST#####|$name|g"  $PIRATEBOX_FOLDER/src/redirect.html.schema >  $WWW_FOLDER/redirect.html
+	sed  "s|#####HOST#####|$name|g"  $PIRATEBOX_FOLDER/src/redirect.html.schema >  $WWW_FOLDER/index.html
         sed "s|HOST=\"$HOST\"|HOST=\"$name\"|" -i  $PIRATEBOX_CONFIG
 }
 
@@ -167,3 +173,6 @@ if [ $2 = "hostname" ] ; then
 	echo "..done"
 fi
 
+if [ $2 = "content" ] ; then
+	create_content_folder
+fi
